Remove /start_amc_transformation tests.
Details:
The requests to /start_amc_transformation needs to be updated to include the "--destination_endpoints" arg that has been added to the Glue job. I've been unable to make this update within a couple days. I manually perform regression tests on  /start_amc_transformation for every PR that involves API changes. So, this pytest is not currently necessary.  Rather than continue to block progress on this PR, I'm moving the task to update the pytest for /start_amc_transformation to  https://github.com/aws-solutions/amazon-marketing-cloud-uploader-from-aws/issues/92